chore: apply dev cleanup
- remove db-related code from the project
- clean up npm scripts
- adjust tailwind config for compatability
