+++ .github/workflows/ci-checks.yml
@@ -37,9 +37,7 @@ jobs:
       - name: Install dependencies
         run: npm ci
 
-      - name: Generate Prisma client # required to generate types for lint checks
-        run: npm run generate
-
+      # Ensure the project builds
       - name: Build project
         run: npm run build
 
@@ -98,9 +96,7 @@ jobs:
       - name: Install dependencies
         run: npm ci
 
-      - name: Generate Prisma client
-        run: npm run generate
-
+# Ensure the project builds
       - name: Build project
         run: npm run build
