+++ package.json
@@ -17,9 +17,7 @@
     "lint": "eslint . --no-cache",
     "prettier:fix": "prettier --write \"**/*.{js,jsx,ts,tsx,json,md}\"",
     "prettier:check": "prettier --check \"**/*.{js,jsx,ts,tsx,json,md}\"",
-    "test": "concurrently \"npm run test -w frontend\" \"npm run test -w backend\"",
-    "migrate": "npm run migrate -w backend",
-    "generate": "npm run generate -w backend"
+    "test": "concurrently \"npm run test -w frontend\" \"npm run test -w backend\""
   },
   "devDependencies": {
     "concurrently": "^9.0.1"
