+++ palette-types/package.json
@@ -4,7 +4,7 @@
   "main": "dist/index.d.ts",
   "scripts": {
     "build": "tsc",
-    "clean": "sudo rm -rf dist && sudo rm -rf node_modules && npm install"
+    "clean": "rm -rf dist && rm -rf node_modules && npm install && npm run build"
   },
   "files": [
     "dist"
